+++ src/main/java/com/ryuu/movieReservation/model/Showtime.java
@@ -8,6 +8,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -37,7 +38,7 @@ public class Showtime {
 
 
     @Column(name="seat_price")
-    private double seatPrice;
+    private BigDecimal seatPrice;
 
     @OneToMany(mappedBy = "showtime", cascade = CascadeType.ALL, orphanRemoval = true)
     @JsonIgnore
